requirements.txt
accelerate==1.8.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
faiss-cpu==1.7.4
fastapi==0.116.0
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
lightning-utilities==0.14.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
pytorch-metric-learning==2.8.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.16
tokenizers==0.21.2
torch==2.7.1
torchmetrics==1.7.4
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.1
typing-inspection==0.4.1
typing_extensions==4.14.1
unicorn==2.1.3
urllib3==2.5.0
uvicorn==0.35.0
